database user creation
users logging in are checked to see if they have a record in the database. If no, record is created for them and saved in the db. If yes, then ignored.
Run npm install mysql in /database.
Download HeidiSQL (mongodb?) to view the database, needs to be set up though. Username and password are in the model-mysql.js file.