DB now accepts valid events
events created are added to the database. note: admin stuff not yet sorted out, need to validate email entries, ensure they are already in the user table before assigning them as a foreign key in the event database. 
Once again, use heidiSQL (mongoDB) to set up a session to the database.